repo_content/nexus_agent/requirements.txt
-e ../nexus_shared
google-adk>=1.3.0
mcp>=1.9.0
litellm
pydantic-settings
rich
click
google-genai
